bot/requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
python-decouple==3.8
pytz==2024.2
pyaml==23.7.0